+++ vite.config.ts
@@ -65,6 +65,7 @@ const ALIASES = {
   '@filetransfer': resolve(__dirname, 'src/filetransfer'),
   '@graph': resolve(__dirname, 'src/graph'),
   '@ftelnetclient': resolve(__dirname, 'src/ftelnetclient'),
+  '@i18n': resolve(__dirname, 'src/i18n'),
 };
 
 export default defineConfig(({ mode }): UserConfig => {